inattro/requirements.txt
asgiref==3.8.1
beautifulsoup4==4.12.3
certifi==2024.8.30
charset-normalizer==3.4.0
Django==5.1.2
djangorestframework==3.15.2
drf-yasg==1.21.8
idna==3.10
inflection==0.5.1
packaging==24.1
pytz==2024.2
PyYAML==6.0.2
requests==2.32.3
soupsieve==2.6
sqlparse==0.5.1
uritemplate==4.1.1
urllib3==2.2.3
